feat(admin-ui): rename to exploration
