+++ packages/admin-ui/pages/exploration/index.tsx
@@ -1,11 +1,16 @@
 import getConfig from 'next/config';
-import { useRouter } from 'next/router';
 import { GetServerSideProps } from 'next';
+import { useRouter } from 'next/router';
+import { Button } from 'antd';
 import { Path } from '@vulcan-sql/admin-ui/utils/enum';
+import { ExploreIcon } from '@vulcan-sql/admin-ui/utils/icons';
 import SiderLayout from '@vulcan-sql/admin-ui/components/layouts/SiderLayout';
+import SelectDataToExploreModal from '@vulcan-sql/admin-ui/components/pages/explore/SelectDataToExploreModal';
 import Background from '@vulcan-sql/admin-ui/components/Background';
+import useModalAction from '@vulcan-sql/admin-ui/hooks/useModalAction';
 
-export default function Explore({ connections }) {
+export default function Exploration({ connections }) {
+  const selectDataToExploreModal = useModalAction();
   const router = useRouter();
 
   // TODO: call API to get real exploration list data
@@ -21,7 +26,7 @@ export default function Explore({ connections }) {
   ];
 
   const onSelect = (selectKeys: string[]) => {
-    router.push(`${Path.Explore}/${selectKeys[0]}`);
+    router.push(`${Path.Exploration}/${selectKeys[0]}`);
   };
 
   return (
@@ -31,7 +36,23 @@ export default function Explore({ connections }) {
       sidebar={{ data, onSelect }}
     >
       <Background />
-      Explore: {router.query.id}
+
+      <div
+        className="d-flex align-center justify-center"
+        style={{ height: '100%' }}
+      >
+        <Button
+          icon={<ExploreIcon className="mr-2" />}
+          onClick={() => selectDataToExploreModal.openModal()}
+        >
+          Start from modeling
+        </Button>
+      </div>
+
+      <SelectDataToExploreModal
+        {...selectDataToExploreModal.state}
+        onClose={selectDataToExploreModal.closeModal}
+      />
     </SiderLayout>
   );
 }
